# app/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
  /auth/register:
    post:
      summary: Register User
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "example@email.com"
                  description: "The user's email address"
                password:
                  type: string
                  format: password
                  example: "yourpassword"
                  description: "The user's password"
                name:
                  type: string
                  format: name
                  example: "yourname"
                  description: "The user's name"
              required:
                - email
                - password
                - name
      responses:
        201:
          description: Success
  /auth/login:
    post:
      summary: User Login
      description: >
        Authenticates the user with their email and password. After a successful login, you'll receive an `access_token`. 
        
        
        **Steps:**
        
        
        1. Copy the `access_token` from the response.
        
        2. Click the 'Authorize' button in the top-right corner.
        
        3. Enter the token in the format `<your_access_token>`
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "example@email.com"
                  description: "The user's email address"
                password:
                  type: string
                  format: password
                  example: "yourpassword"
                  description: "The user's password"
              required:
                - email
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Login successful"
        '400':
          description: Missing request data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Required data is missing."
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "Invalid credentials."
  /auth/refresh:
    post:
      summary: Refresh access token
      description: This endpoint generates a new access token using a valid refresh token.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token provided during login.
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        200:
          description: Tokens have been refreshed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Tokens have been refreshed.
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        401:
          description: Invalid or expired refresh token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Invalid refresh token.
  /auth/profile:
    put:
      summary: Update user profile
      description: Updates the user's password and/or name. Requires JWT authentication.
      tags:
        - Auth
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: New password (Base64 encoded).
                name:
                  type: string
                  description: New name of the user.
              example:
                password: new_password123
                name: new_name
      responses:
        '200':
          description: Profile successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: User information has been successfully updated.
        '400':
          description: Missing or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Invalid refresh token
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
  /auth/info:
    get:
      summary: Get User Info
      description: Retrieves the profile information of the currently authenticated user.
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: example@email.com
                      name:
                        type: string
                        example: John Doe
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
                      last_login:
                        type: string
                        format: date-time
                        example: 2024-12-01T11:00:00Z
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: User not found.
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
  /auth/delete:
    delete:
      summary: Delete User Account
      description: Deletes the account of the currently authenticated user.
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful deletion
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Account has been successfully deleted.
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT